requirements.txt
pandas
matplotlib
requests
notebook
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
Flask
Flask-Cors
asgiref
mlflow==2.7.0
dagshub
transformers
transformers[sentencepiece]
accelerate
accelerate[sagemaker]
datasets
sacrebleu
rouge_score
bert_score
py7zr
boto3
botocore
langchain
spacy
torchvision
pyaudio
